Re-working progress interface. closes #394
git-svn-id: file:///storage/svn/bagit@1674 b77ddd6a-a23c-0410-94b0-b3c43762c485

+++ bagit/trunk/src/main/java/gov/loc/repository/bagit/transfer/BagFetcher.java
@@ -92,7 +92,7 @@ public final class BagFetcher implements ActiveCancellable, ProgressListenable
     	this.progressListeners.remove(progressListener);
     }
     
-    private void progress(String activity, String item, int count, int total)
+    private void progress(String activity, String item, long count, long total)
     {
     	for (ProgressListener listener : this.progressListeners)
     	{
