+++ bagit/trunk/src/main/java/gov/loc/repository/bagit/writer/impl/AbstractWriter.java
@@ -51,7 +51,7 @@ public abstract class AbstractWriter extends AbstractBagVisitor implements Write
 		this.progressListeners.remove(progressListener);
 	}
 	
-	protected void progress(String activity, String item, int count, int total)
+	protected void progress(String activity, String item, long count, long total)
 	{
 		for (ProgressListener listener : this.progressListeners)
 		{
